pyproject.toml
[project]
name = "safevision"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.124.4",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.12.0",
    "uvicorn>=0.40.0",
    "opencv-python>=4.12.0.88",
    "tensorflow>=2.20.0",
    "sqlalchemy>=2.0.45",
    "psycopg2-binary>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt[crypto,cryptography]>=2.10.1",
    "python-multipart>=0.0.21",
]
